refactor: (#44) jwt, cors 보안설정 개선
