+++ Backend/src/main/java/com/demo/nimn/config/cors/DevCorsConfig.java
@@ -1,22 +1,20 @@
-package com.demo.nimn.config;
+package com.demo.nimn.config.cors;
 
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 @Configuration
-public class WebConfig implements WebMvcConfigurer {
+@Profile("dev") // 개발 환경에서만 활성화
+public class DevCorsConfig implements WebMvcConfigurer {
     @Override
     public void addCorsMappings(CorsRegistry registry) {
         registry.addMapping("/**")
-                .allowedOrigins(
-                        "http://3.23.4.160",
-                        "http://127.0.0.1:3000",
-                        "http://nimn.store",
-                        "https://nimn.store",
-                        "https://nimn.netlify.app")
-                .allowedMethods("*")
+                .allowedOrigins("http://localhost:3000", "http://127.0.0.1:3000")
+                .allowedMethods("GET", "POST", "PUT", "DELETE", "OPTIONS")
                 .allowedHeaders("*")
                 .allowCredentials(true);
     }
 }
+
