+++ Backend/src/main/java/com/demo/nimn/config/cors/WebConfig.java
@@ -1,0 +1,19 @@
+package com.demo.nimn.config.cors;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+@Profile("prod") // 운영 환경에서만 활성화
+public class WebConfig implements WebMvcConfigurer {
+    @Override
+    public void addCorsMappings(CorsRegistry registry) {
+        registry.addMapping("/**")
+                .allowedOrigins("https://nimn.store", "https://nimn.netlify.app")
+                .allowedMethods("GET", "POST")
+                .allowedHeaders("Content-Type", "Authorization")
+                .allowCredentials(true);
+    }
+}
