+++ Backend/src/main/java/com/demo/nimn/filter/JWTFilter.java
@@ -30,7 +30,7 @@ public class JWTFilter extends OncePerRequestFilter {
         //Authorization 헤더 검증
         if (authorization == null || !authorization.startsWith("Bearer ")) {
 
-            System.out.println("token null");
+
             filterChain.doFilter(request, response);
 
             //조건이 해당되면 메소드 종료 (필수)
@@ -44,7 +44,7 @@ public class JWTFilter extends OncePerRequestFilter {
         //토큰 소멸 시간 검증
         if (jwtUtil.isExpired(token)) {
 
-            System.out.println("token expired");
+
             filterChain.doFilter(request, response);
 
             //조건이 해당되면 메소드 종료 (필수)
