+++ Backend/src/main/resources/application.properties
@@ -20,11 +20,10 @@ import.secretKey=${import_secretKey}
 
 spring.jpa.properties.hibernate.jdbc.time_zone=Asia/Seoul
 
-#spring.jwt.secret=${JWT_SECRET}
-spring.jwt.secret=bXlTZWNyZXRLZXlGb3JKV1RUb2tlbkdlbmVyYXRpb24yMDI0IUAjJCU=
+spring.jwt.secret=${JWT_SECRET}
 
 
-#email ??
+#email
 spring.mail.host=smtp.gmail.com
 spring.mail.port=587
 spring.mail.username=${GOOGLE_EMAIL}
